Move Target Actor launch to mod target_actor
